requirements.txt
flask
flask-socketio
requests
opencv-python
torch
torchvision
Pillow
matplotlib
pyserial
